// frontend/src/types/predictionData.ts
// Define an interface for the data structure
export interface PredictionData {
    Precio_promedio: number;
    Costos: number;
    Rotacion: number;
    Marketing: number;
    Ingresos_totales: number;
    Costos_operativos: number;
    Precio_competencia: number;
    Demanda_sectorial: number;
    Tasa_CUP_USD: number;
    Ventas: number;
    Beneficio_neto: number;
    Ingresos_totales_final: number;
}